backend/worker/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
readability-lxml==0.8.1
boto3==1.28.63
python-dotenv==1.0.0
pytest==7.4.3
yt-dlp==2023.11.16
